+++ src/main/java/View/View.java
@@ -1,9 +1,0 @@
-package View;
-
-public class View {
-    // TODO Transfer all the views from Controller class here
-
-    public static void LoginView() {
-        // TODO: Implement this method
-    }
-}