+++ src/main/java/controller/Controller.java
@@ -1,11 +1,12 @@
-package Controller;
+package controller;
 
-import java.sql.SQLOutput;
+//import java.sql.SQLOutput;
 import java.util.*;
 
-import Use_Cases.*;
-import Entities.*;
-import Data.*;
+import use_cases.*;
+import entities.*;
+import data.*;
+import view.*;
 
 public class Controller {
 
@@ -13,39 +14,20 @@ public class Controller {
     private static boolean isLoggedIn = Boolean.FALSE;
 
     // TODO: Replace the following dummy variable for app name
-    private static String appName = "[APP NAME]";
-    public static String[] mainMenuOptions = {"Sign in to my account", "Create a new account"};
+    public static String appName = "[APP NAME]";
 
     private static Scanner sc = new Scanner(System.in);
 
-    public static void signUp() {
-        String[] outputs = {"Full Name (*): ", "Email (*): ", "Phone: "};
-        List<String> inputs = new ArrayList<>();
 
-        System.out.println("Please enter the following information and press enter. Fields marked (*) are required.");
-        for (String output: outputs) {
-            System.out.println(output);
-            String input = sc.nextLine();
-            inputs.add(input);
-        }
-
-        currentUser = new User(inputs.get(0),0,inputs.get(2));
-
-        System.out.println("Thanks for signing up! Your account has been successfully created, " + inputs.get(0) + ".");
-    }
 
 
     public static void main(String[] args) {
         Data.initializeData();
-        currentUser = (User) Data.USERS.get(0); // TODO Change this to authenticated user in submission version
+        do {
+            View.menuView();
+        } while (!isLoggedIn);
 
-        isLoggedIn = Boolean.TRUE;
-        System.out.println("Welcome to " + appName);
-        for (int i = 1; i <= mainMenuOptions.length; i++) {
-            System.out.println(i + ") " + mainMenuOptions[i - 1]);
-        }
         System.out.println("\n");
-//        signUp();
 
         while (isLoggedIn) {
             System.out.println("""
@@ -55,19 +37,21 @@ public class Controller {
                     3. Check balance
                     4. Update Profile [Coming soon]
                     5. Create a new group
-                    6. Log out""");
+                    6. View expenses
+                    7. Log out""");
             String input = sc.nextLine();
             switch (input) {
 //                case "1" -> GroupManager.create_temp();
                 case "1" -> createExpenseView();
                 case "2" -> {
-                    StringBuilder lst = GroupManager.show_group(currentUser);
+                    StringBuilder lst = ExpenseManager.show_group(currentUser);
                     System.out.println(lst);
                 }
                 case "3" -> System.out.println("Your balance is: $" + currentUser.getBalance());
                 case "4" -> UserManager.updateProfile(currentUser);
-                case "5" -> createGroupView();
-                case "6" -> {
+                case "5" -> View.createGroupView();
+                case "6" -> System.out.println(UserManager.getExpenses(currentUser));
+                case "7" -> {
                     currentUser = null;
                     isLoggedIn = Boolean.FALSE;
                     System.out.println("Goodbye. Have a nice day!");
@@ -82,6 +66,7 @@ public class Controller {
     public static void createExpenseView() {
         double amount;
         List<String> people = new ArrayList<>();
+        people.add(currentUser.getEmail());
 
         System.out.println("Enter amount: ");
         amount = Float.parseFloat(sc.nextLine());
@@ -94,8 +79,24 @@ public class Controller {
 
         // GROUP EXPENSE
         if (input.equals("y") || input.equals("Y")) {
-            // Expense.createGroupExpense();
-            System.out.println("Group expenses are not currently supported.");
+            StringBuilder lst = ExpenseManager.show_group(currentUser);
+            System.out.println(lst);
+            System.out.println("Enter group name: ");
+            String groupName = sc.nextLine();
+            try {
+                // TODO Implement this as Group.findGroup rather than directly
+                for (Group group: Data.groups) {
+                    if (group.getGroupName().equals(groupName)) {
+                        if (Expense.createGroupExpense("", amount, currentUser.getUUID(), group)) {
+                            System.out.println("Successfully added to your expenses.");
+                        }
+                        break;
+                    }
+                }
+            } catch (Exception e) {
+                System.out.println("There was an error finding your group in our database.");
+            }
+//            System.out.println("Group expenses are not currently supported.");
         }
 
         // NOT A GROUP EXPENSE
@@ -122,7 +123,8 @@ public class Controller {
             } while (addMorePeople);
             if (Expense.createExpense(amount, currentUser.getUUID(), people)) {
                 System.out.println("Expense has been successfully created!");
-                System.out.println(Data.EXPENSES);
+                System.out.println(Data.expenses);
+                System.out.println(currentUser.expenses.get(0));
             }
         } else {
             // TODO: Handle this
@@ -131,65 +133,16 @@ public class Controller {
 
     }
 
-    private static void authenticateUser() {
+    public static void authenticateUser(User user) {
         // TODO: Implement this method
+        currentUser = user; // TODO Set it as indexOf user in Data.USER insetead of directly assigning user object
         isLoggedIn = Boolean.TRUE;
+        System.out.println("Welcome back, " + currentUser.getName() + "!");
+//        view.dashboardView();
     }
 
 
-    /** View to create new groups
-     * @returns 0: if user is not authenticated, the view doesn't allow new group to be created.
-     */
-    public static int createGroupView() {
-        if (!isLoggedIn) {
-            System.out.println("Error: You must be authenticated to create a new group.");
-            return 0;
-        }
 
-
-        String gName;
-        List<String> members = new ArrayList<>();
-        members.add(currentUser.getEmail());
-
-
-        //
-        System.out.println("Enter name of the group: ");
-        gName = sc.nextLine();
-
-        boolean addAnotherMember = false;
-
-        System.out.println("ADD GROUP MEMBERS:");
-        System.out.println("You will now be asked to add group members. " +
-                "Press enter if you don't want to add any member");
-
-        do {
-            System.out.println("Enter email of member " + members.size() + ": ");
-            String member = sc.nextLine();
-            if (member.equals("")) { continue; }
-            members.add(member);
-
-            System.out.println("Would you like to add more members? (y/n)");
-
-            if (sc.nextLine().equals("y")) {
-                addAnotherMember = Boolean.TRUE;
-            } else {
-                addAnotherMember = Boolean.FALSE;
-            }
-        } while (addAnotherMember);
-
-        Group g1 = new Group(
-                gName, members, new ArrayList<Expense>(), "Edit group description in Manage Group."
-        );
-
-        Data.GROUPS.add(g1);
-
-        /* For testing the code */
-        System.out.println(Data.GROUPS);
-        System.out.println(Data.GROUPS.get(1).getGroupName());
-        /* */
-
-        return 1;
-    }
 
     /**
      * Returns user's unique identifier through email
@@ -197,7 +150,7 @@ public class Controller {
      * @return UUID of user is user with given email exists in Data.USERS, "0" otherwise
      */
     public static String getUUID(String email) {
-        for (Person person: Data.USERS) {
+        for (Person person: Data.users) {
             try {
                 User user = (User) person;
                 if (user.getEmail().equals(email)) {
@@ -207,5 +160,32 @@ public class Controller {
         }
 
         return "0";
+    }
+
+    public static User getUser(String email) {
+        try {
+            for (Person person: Data.users) {
+                if (person.getEmail().equals((email))) {
+                    return (User) person;
+                }
+            }
+        } catch (Exception ignored) { }
+        return null;
+    }
+
+    /**
+     * Function
+     * @return true, if user is logged in. False otherwise.
+     */
+    public static boolean getUserStatus() {
+        return isLoggedIn;
+    }
+
+    /**
+     * Function
+     * @return current user
+     */
+    public static User getCurrentUser() {
+        return currentUser;
     }
 }